Reorganize events code into more readable files
