+++ js/view/components/graph/events/index.js
@@ -1,0 +1,9 @@
+import bindHover from './hover';
+import bindExpandCollapse from './expandCollapse';
+
+const bindEvents = (cy) => {
+  bindHover(cy);
+  bindExpandCollapse(cy);
+};
+
+export default bindEvents;
